fix id type and add log

+++ app/db/models/base.py
@@ -1,13 +1,13 @@
 from app.db.db import Base
 
-from sqlalchemy import Column, Integer, DateTime
+from sqlalchemy import Column, BigInteger, DateTime
 from datetime import datetime, timezone
 
 
 class BaseModel(Base):
     __abstract__ = True
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc), nullable=False)
     updated_at = Column(DateTime, default=lambda: datetime.now(timezone.utc), nullable=False)
